chore: implement arbitrary for `Config` (#3447)

+++ src/rpc/mod.rs
@@ -57,7 +57,7 @@ where
     use sync_api::*;
     use wallet_api::*;
 
-    let block_delay = state.state_manager.chain_config().block_delay_secs;
+    let block_delay = state.state_manager.chain_config().block_delay_secs as u64;
     let rpc_server = Arc::new(
         Server::new()
             .with_data(Data(state))
